requirement.txt
asgiref==3.2.7
authy==2.2.5
autopep8==1.5
certifi==2019.11.28
chardet==3.0.4
Django==3.0.4
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
idna==2.9
phonenumbers==8.12.0
pycodestyle==2.5.0
PyJWT==1.7.1
python-decouple==3.3
pytz==2019.3
requests==2.23.0
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
